Refactor and cleanup the code + add .nvmrc
