+++ .nvmrc
@@ -1,0 +1,1 @@
+16.19.0
